requirements.txt
# Add python libraries and their respective versions here
# pip list
# pip install --upgrade pip
# brew install ffmpeg

# Required python libraries to run project
allure-pytest==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==4.2.0
websocket-client==1.8.0
attrs==23.2.0
flake8==7.0.0
flake8-builtins==2.2.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.2
pluggy==1.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.0.0
setuptools==69.0.3
wheel==0.42.0

# Required python libraries for automated testing of UI
selenium==4.17.2
webdriver-manager==4.0.1
chromedriver_autoinstaller==0.6.4

# Additional python libraries required for validating test cases
pandas==2.2.0
tabulate==0.9.0
openpyxl==3.1.2
pyarrow==14.0.2
pytesseract==0.3.10
requests==2.31.0
pillow==10.3.0
cryptography==42.0.8